\name{sinib-package}
\alias{sinib-package}
\alias{sinib}
\docType{package}
\title{
\packageTitle{sinib}
}
\description{
\packageDescription{sinib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sinib}
\packageIndices{sinib}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{sinib}

Maintainer: \packageMaintainer{sinib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
